@@ -10,7 +10,7 @@ import androidx.core.content.ContextCompat
 import coil.load
 import com.anteprocess.food.R
 import com.anteprocess.food.data.util.Constants.Companion.RECIPE_RESULT
-import kotlinx.android.synthetic.main.fragment_over_view.view.*
+import com.anteprocess.food.databinding.FragmentOverViewBinding
 import org.jsoup.Jsoup
 /**
  * A simple [Fragment] subclass.
@@ -19,58 +19,67 @@ import org.jsoup.Jsoup
  */
 class OverViewFragment : Fragment() {
 
+    private var _binding : FragmentOverViewBinding? = null
+    private val bindings get() = _binding!!
+
     override fun onCreateView(
         inflater: LayoutInflater, container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View? {
         // Inflate the layout for this fragment
-        val view =  inflater.inflate(R.layout.fragment_over_view, container, false)
+        _binding =  FragmentOverViewBinding.inflate(inflater, container, false)
 
         val args = arguments
         val myBundle: Result? = args?.getParcelable(RECIPE_RESULT)
-        view.main_imageView.load(myBundle?.image)
-        view.title_textView.text = myBundle?.title
-        view.likes_textView.text = myBundle?.aggregateLikes.toString()
-        view.time_textView.text = myBundle?.readyInMinutes.toString()
+        bindings.mainImageView.load(myBundle?.image)
+
+        bindings.titleTextView.text = myBundle?.title
+        bindings.likesTextView.text = myBundle?.aggregateLikes.toString()
+        bindings.timeTextView.text = myBundle?.readyInMinutes.toString()
 
         myBundle?.summary.let {
             val summary = Jsoup.parse(it).text()
-            view.summary_textView.text = summary
+            bindings.summaryTextView.text = summary
         }
 
         if (myBundle?.vegetarian == true) {
-            view.vegetarian_imageView.setColorFilter(ContextCompat.getColor(requireContext(), R.color.green ))
-            view.vegetarian_textView.setTextColor(ContextCompat.getColor(requireContext(), R.color.green ))
+            bindings.vegetarianImageView.setColorFilter(ContextCompat.getColor(requireContext(), R.color.green ))
+            bindings.vegetarianTextView.setTextColor(ContextCompat.getColor(requireContext(), R.color.green ))
         }
 
 
         if (myBundle?.vegan == true) {
-            view.vegan_imageView.setColorFilter(ContextCompat.getColor(requireContext(), R.color.green ))
-            view.vegan_textView.setTextColor(ContextCompat.getColor(requireContext(), R.color.green ))
+            bindings.veganImageView.setColorFilter(ContextCompat.getColor(requireContext(), R.color.green ))
+            bindings.veganTextView.setTextColor(ContextCompat.getColor(requireContext(), R.color.green ))
         }
 
         if (myBundle?.glutenFree == true) {
-            view.gluten_free_imageView.setColorFilter(ContextCompat.getColor(requireContext(), R.color.green ))
-            view.gluten_free_textView.setTextColor(ContextCompat.getColor(requireContext(), R.color.green ))
+            bindings.glutenFreeImageView.setColorFilter(ContextCompat.getColor(requireContext(), R.color.green ))
+            bindings.glutenFreeTextView.setTextColor(ContextCompat.getColor(requireContext(), R.color.green ))
         }
 
         if (myBundle?.dairyFree == true) {
-            view.dairy_free_imageView.setColorFilter(ContextCompat.getColor(requireContext(), R.color.green ))
-            view.dairy_free_textView.setTextColor(ContextCompat.getColor(requireContext(), R.color.green ))
+            bindings.dairyFreeImageView.setColorFilter(ContextCompat.getColor(requireContext(), R.color.green ))
+            bindings.dairyFreeTextView.setTextColor(ContextCompat.getColor(requireContext(), R.color.green ))
         }
 
         if (myBundle?.veryHealthy == true) {
-            view.healthy_imageView.setColorFilter(ContextCompat.getColor(requireContext(), R.color.green ))
-            view.healthy_textView.setTextColor(ContextCompat.getColor(requireContext(), R.color.green ))
+            bindings.healthyImageView.setColorFilter(ContextCompat.getColor(requireContext(), R.color.green ))
+            bindings.healthyTextView.setTextColor(ContextCompat.getColor(requireContext(), R.color.green ))
         }
 
         if (myBundle?.cheap == true) {
-            view.cheap_imageView.setColorFilter(ContextCompat.getColor(requireContext(), R.color.green ))
-            view.cheap_textView.setTextColor(ContextCompat.getColor(requireContext(), R.color.green ))
+            bindings.cheapImageView.setColorFilter(ContextCompat.getColor(requireContext(), R.color.green ))
+            bindings.cheapTextView.setTextColor(ContextCompat.getColor(requireContext(), R.color.green ))
         }
 
 
-        return view
+        return bindings.root
+    }
+
+    override fun onDestroyView() {
+        super.onDestroyView()
+        _binding = null
     }
 
 }
