@@ -9,30 +9,30 @@ import coil.load
 import com.anteprocess.food.R
 import com.anteprocess.food.data.util.Constants.Companion.BASE_IMAGE_URL
 import com.anteprocess.food.data.util.RecipesDiffUtil
+import com.anteprocess.food.databinding.IngredientsRowLayoutBinding
 import com.anteprocess.food.models.ExtendedIngredient
-import kotlinx.android.synthetic.main.ingredients_row_layout.view.*
 import java.util.*
 
 class IngredientsAdapter:  RecyclerView.Adapter<IngredientsAdapter.MyViewHolder>() {
 
     private var ingredientsList = emptyList<ExtendedIngredient>()
 
-    class MyViewHolder(itemView: View): RecyclerView.ViewHolder(itemView)
+    class MyViewHolder(val binding: IngredientsRowLayoutBinding): RecyclerView.ViewHolder(binding.root)
 
     override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): MyViewHolder {
-        return MyViewHolder(LayoutInflater.from(parent.context).inflate(R.layout.ingredients_row_layout, parent, false))
+        return MyViewHolder(IngredientsRowLayoutBinding.inflate(LayoutInflater.from(parent.context), parent, false))
     }
 
     override fun onBindViewHolder(holder: MyViewHolder, position: Int) {
-        holder.itemView.ingredient_imageView.load(BASE_IMAGE_URL + ingredientsList[position].image) {
+        holder.binding.ingredientImageView.load(BASE_IMAGE_URL + ingredientsList[position].image) {
             crossfade(600)
             error(R.drawable.ic_error_placeholder)
         }
-        holder.itemView.ingredient_name.text = ingredientsList[position].name?.capitalize(Locale.ROOT)
-        holder.itemView.ingredient_amount.text = ingredientsList[position].amount.toString()
-        holder.itemView.ingredient_unit.text = ingredientsList[position].unit
-        holder.itemView.ingredient_consistency.text = ingredientsList[position].consistency
-        holder.itemView.ingredient_original.text = ingredientsList[position].original
+        holder.binding.ingredientName.text = ingredientsList[position].name?.capitalize(Locale.ROOT)
+        holder.binding.ingredientAmount.text = ingredientsList[position].amount.toString()
+        holder.binding.ingredientUnit.text = ingredientsList[position].unit
+        holder.binding.ingredientConsistency.text = ingredientsList[position].consistency
+        holder.binding.ingredientOriginal.text = ingredientsList[position].original
     }
 
     override fun getItemCount(): Int {
