@@ -2,25 +2,22 @@ package com.anteprocess.food.adapters
 
 import android.os.Bundle
 import androidx.fragment.app.Fragment
+import androidx.fragment.app.FragmentActivity
 import androidx.fragment.app.FragmentManager
 import androidx.fragment.app.FragmentPagerAdapter
+import androidx.viewpager2.adapter.FragmentStateAdapter
 
 class PagerAdapter(
     private val resultBundle: Bundle,
     private val fragments: ArrayList<Fragment>,
-    private val title: ArrayList<String>,
-    fm: FragmentManager
-) : FragmentPagerAdapter(fm, BEHAVIOR_RESUME_ONLY_CURRENT_FRAGMENT) {
-    override fun getCount(): Int {
+    fragmentActivity: FragmentActivity
+) : FragmentStateAdapter(fragmentActivity) {
+    override fun getItemCount(): Int {
         return fragments.size
     }
 
-    override fun getItem(position: Int): Fragment {
+    override fun createFragment(position: Int): Fragment {
         fragments[position].arguments = resultBundle // passing the args to the frag
         return fragments[position]
-    }
-
-    override fun getPageTitle(position: Int): CharSequence? {
-        return title[position]
     }
 }
