@@ -3,47 +3,33 @@ package com.anteprocess.food.bindingadapters
 import android.view.View
 import android.widget.ImageView
 import android.widget.TextView
+import androidx.core.view.isVisible
 import androidx.databinding.BindingAdapter
 import com.anteprocess.food.data.database.entities.RecipesEntity
 import com.anteprocess.food.data.util.NetworkResult
 import com.anteprocess.food.models.FoodRecipe
 
 class RecipesBinding {
+
     companion object {
 
-        // Show error placeholder if there are no data or has an error
-
-        @BindingAdapter("readApiResponse", "readDatabase", requireAll = true)
+        @BindingAdapter("readApiResponse","readDatabase",requireAll = true)
         @JvmStatic
-        fun errorImageViewVisibility(
-            imageView: ImageView,
+        fun handleReadDataErrors(
+            view: View,
             apiResponse: NetworkResult<FoodRecipe>?,
             database: List<RecipesEntity>?
-        ) {
-            if (apiResponse is NetworkResult.Error && database.isNullOrEmpty()){
-                imageView.visibility = View.VISIBLE
-            } else if (apiResponse is NetworkResult.Loading) {
-                imageView.visibility = View.INVISIBLE
-            } else if (apiResponse is NetworkResult.Success) {
-                imageView.visibility = View.INVISIBLE
-            }
-        }
-
-        @BindingAdapter("readApiResponse2", "readDatabase2", requireAll = true)
-        @JvmStatic
-        fun errorTextViewVisibility(
-            textView: TextView,
-            apiResponse: NetworkResult<FoodRecipe>?,
-            database: List<RecipesEntity>?
-        ) {
-            if (apiResponse is NetworkResult.Error && database.isNullOrEmpty()){
-                textView.visibility = View.VISIBLE
-                textView.text = apiResponse.message.toString()
-            } else if (apiResponse is NetworkResult.Loading) {
-                textView.visibility = View.INVISIBLE
-            } else if (apiResponse is NetworkResult.Success) {
-                textView.visibility = View.INVISIBLE
+        ){
+            when (view){
+                is ImageView ->{
+                    view.isVisible = apiResponse is NetworkResult.Error && database.isNullOrEmpty()
+                }
+                is TextView ->{
+                    view.isVisible = apiResponse is NetworkResult.Error && database.isNullOrEmpty()
+                    view.text = apiResponse?.message.toString()
+                }
             }
         }
     }
+
 }
