@@ -51,17 +51,17 @@ class RecipesRowBinding {
             }
         }
 
-        @BindingAdapter("setNumberOfLikes")
-        @JvmStatic
-        fun setNumberOfLikes(textView: TextView, likes: Int){
-            textView.text = likes.toString()
-        }
-
-        @BindingAdapter("setNumberOfMinutes")
-        @JvmStatic
-        fun setNumberOfMinutes(textView: TextView, minutes: Int){
-            textView.text = minutes.toString()
-        }
+//        @BindingAdapter("setNumberOfLikes")
+//        @JvmStatic
+//        fun setNumberOfLikes(textView: TextView, likes: Int){
+//            textView.text = likes.toString()
+//        }
+//
+//        @BindingAdapter("setNumberOfMinutes")
+//        @JvmStatic
+//        fun setNumberOfMinutes(textView: TextView, minutes: Int){
+//            textView.text = minutes.toString()
+//        }
 
         @BindingAdapter("applyVeganColor")
         @JvmStatic
