@@ -2,20 +2,21 @@ package com.anteprocess.food.models
 
 import android.os.Parcelable
 import com.google.gson.annotations.SerializedName
-import kotlinx.android.parcel.Parcelize
+import kotlinx.parcelize.Parcelize
+
 
 @Parcelize
 data class ExtendedIngredient(
     @SerializedName("amount")
-    val amount: Double,
+    val amount: Double = 0.0,
     @SerializedName("consistency")
-    val consistency: String,
+    val consistency: String?,
     @SerializedName("image")
-    val image: String,
+    val image: String?,
     @SerializedName("name")
-    val name: String,
+    val name: String?,
     @SerializedName("original")
-    val original: String,
+    val original: String?,
     @SerializedName("unit")
-    val unit: String
+    val unit: String?
 ): Parcelable
