@@ -2,7 +2,6 @@ package com.anteprocess.food.viewmodels
 
 import android.app.Application
 import android.widget.Toast
-import androidx.hilt.lifecycle.ViewModelInject
 import androidx.lifecycle.AndroidViewModel
 import androidx.lifecycle.asLiveData
 import androidx.lifecycle.viewModelScope
@@ -19,11 +18,14 @@ import com.anteprocess.food.data.util.Constants.Companion.QUERY_TYPE
 import com.anteprocess.food.data.util.Constants.Companion.QUERY_FILL_INGREDIENTS
 import com.anteprocess.food.data.util.Constants.Companion.QUERY_NUMBER
 import com.anteprocess.food.data.util.Constants.Companion.QUERY_SEARCH
+import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.flow.collect
 import kotlinx.coroutines.launch
+import javax.inject.Inject
 
-class RecipesViewModel @ViewModelInject constructor(
+@HiltViewModel
+class RecipesViewModel @Inject constructor(
     application: Application,
     private val dataStoreRepository: DataStoreRepository
 ) : AndroidViewModel(application) {
